requirements.txt
ccxt>=4.4.0
pandas>=2.0.0
pandas-ta>=0.3.14b
alpaca-py>=0.30.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytz>=2024.1
